+++ src/machine_learning/linear_discriminant_analysis.rs
@@ -1,0 +1,1 @@
+pub use crate::utility::linear_discriminant_analysis::*;