+++ src/utility_test.rs
@@ -1,3 +1,4 @@
 mod principal_component_analysis_test;
 mod train_test_split_test;
 mod kernel_pca_test;
+mod linear_discriminant_analysis_test;